fix: more refactoring for auth
