+++ app/routes/__init__.py
@@ -1,1 +1,4 @@
-# Routes package
+# Import blueprints for easy access
+from . import users, health, items
+
+__all__ = ['users', 'health', 'items']
