+++ app/routes/items.py
@@ -1,1 +1,6 @@
-# Item CRUD endpoints
+from flask import Blueprint
+
+bp = Blueprint('items', __name__)
+
+# Note: /api/items route remains in app/__init__.py for now.
+# Future item CRUD operations should be added here.
