+++ app/templates/base.html
@@ -20,10 +20,10 @@
         <div>
             <a href="{{ url_for('main') }}">Map</a>
             {% if current_user.is_authenticated %}
-                <a href="{{ url_for('logout') }}">Logout ({{ current_user.username }})</a>
+                <a href="{{ url_for('users.logout') }}">Logout ({{ current_user.username }})</a>
             {% else %}
-                <a href="{{ url_for('login') }}">Login</a>
-                <a href="{{ url_for('register') }}">Register</a>
+                <a href="{{ url_for('users.login') }}">Login</a>
+                <a href="{{ url_for('users.register') }}">Register</a>
             {% endif %}
         </div>
     </nav>
