+++ app/templates/register.html
@@ -26,10 +26,7 @@
     </form>
 
     <div style="margin-top: 20px;">
-        Already have an account? <a href="{{ url_for('login') }}" style="color: #444;">Login here</a>
+        Already have an account? <a href="{{ url_for('users.login') }}" style="color: #444;">Login here</a>
     </div>
 </div>
-
-<style>
-</style>
 {% endblock %}
